AgenteCPI_para_github/requirements.txt
requests==2.32.3
python-dotenv==1.0.1
anthropic>=0.40.0
mcp>=1.0.0
flask>=3.0.0
python-docx>=1.1.0
chromadb>=0.5.0
